requirements.txt
beautifulsoup4==4.11.2
boto3==1.26.85
Flask==2.2.3
Flask_Cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2
